refactor(ui): extract shared useDismissable popover hook
