+++ app/frontend/components/share_popover.tsx
@@ -1,6 +1,7 @@
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { useCallback, useRef, useState } from 'react'
 import { createPortal } from 'react-dom'
 import { useMediaQuery } from '../lib/use_media_query'
+import { useDismissable } from '../lib/use_dismissable'
 import { ThemePicker } from './theme_picker'
 
 /** Share is two audiences, one URL: humans get the editor, agents fetching the
@@ -30,23 +31,7 @@ export function SharePopover({ agentsActive }: { agentsActive: number }) {
     })
   }, [])
 
-  useEffect(() => {
-    if (!open) return
-    const close = (e: MouseEvent) => {
-      const target = e.target as Node
-      if (rootRef.current?.contains(target) || popoverRef.current?.contains(target)) return
-      setOpen(false)
-    }
-    const esc = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') setOpen(false)
-    }
-    document.addEventListener('mousedown', close)
-    document.addEventListener('keydown', esc)
-    return () => {
-      document.removeEventListener('mousedown', close)
-      document.removeEventListener('keydown', esc)
-    }
-  }, [open])
+  useDismissable(open, () => setOpen(false), [rootRef, popoverRef])
 
   const popover = (
     <div
